+++ sample/config.js
@@ -41,9 +41,9 @@ System.config({
     "css": "github:systemjs/plugin-css@0.1.20",
     "font-awesome": "npm:font-awesome@4.7.0",
     "google/code-prettify": "github:google/code-prettify@master",
-    "jquery": "npm:jquery@2.2.4",
+    "jquery": "npm:jquery@3.2.1",
     "json": "github:systemjs/plugin-json@0.1.2",
-    "materialize": "npm:materialize-css@0.98.2",
+    "materialize": "npm:materialize-css@0.100.1",
     "nprogress": "github:rstacruz/nprogress@0.2.0",
     "prettify": "npm:prettify@0.1.7",
     "prism": "github:PrismJS/prism@1.3.0",
@@ -87,9 +87,6 @@ System.config({
     },
     "github:jspm/nodelibs-vm@0.1.0": {
       "vm-browserify": "npm:vm-browserify@0.0.4"
-    },
-    "github:jspm/nodelibs-zlib@0.1.0": {
-      "browserify-zlib": "npm:browserify-zlib@0.1.4"
     },
     "github:rstacruz/nprogress@0.2.0": {
       "css": "github:systemjs/plugin-css@0.1.20"
@@ -230,27 +227,8 @@ System.config({
     "npm:babel-runtime@5.8.38": {
       "process": "github:jspm/nodelibs-process@0.1.2"
     },
-    "npm:block-stream@0.0.9": {
-      "assert": "github:jspm/nodelibs-assert@0.1.0",
-      "buffer": "github:jspm/nodelibs-buffer@0.1.1",
-      "inherits": "npm:inherits@2.0.1",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "stream": "github:jspm/nodelibs-stream@0.1.0"
-    },
-    "npm:brace-expansion@1.1.7": {
-      "balanced-match": "npm:balanced-match@0.4.2",
-      "concat-map": "npm:concat-map@0.0.1"
-    },
-    "npm:browserify-zlib@0.1.4": {
-      "assert": "github:jspm/nodelibs-assert@0.1.0",
-      "buffer": "github:jspm/nodelibs-buffer@0.1.1",
-      "pako": "npm:pako@0.2.9",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "readable-stream": "npm:readable-stream@2.2.11",
-      "util": "github:jspm/nodelibs-util@0.1.0"
-    },
     "npm:buffer@5.0.6": {
-      "base64-js": "npm:base64-js@1.2.0",
+      "base64-js": "npm:base64-js@1.2.1",
       "ieee754": "npm:ieee754@1.1.8"
     },
     "npm:builtin-modules@1.1.1": {
@@ -317,42 +295,11 @@ System.config({
     "npm:font-awesome@4.7.0": {
       "css": "github:systemjs/plugin-css@0.1.20"
     },
-    "npm:fs.realpath@1.0.0": {
-      "fs": "github:jspm/nodelibs-fs@0.1.2",
-      "path": "github:jspm/nodelibs-path@0.1.0",
-      "process": "github:jspm/nodelibs-process@0.1.2"
-    },
-    "npm:fstream@1.0.11": {
-      "assert": "github:jspm/nodelibs-assert@0.1.0",
-      "buffer": "github:jspm/nodelibs-buffer@0.1.1",
-      "fs": "github:jspm/nodelibs-fs@0.1.2",
-      "graceful-fs": "npm:graceful-fs@4.1.11",
-      "inherits": "npm:inherits@2.0.1",
-      "mkdirp": "npm:mkdirp@0.3.5",
-      "path": "github:jspm/nodelibs-path@0.1.0",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "rimraf": "npm:rimraf@2.6.1",
-      "stream": "github:jspm/nodelibs-stream@0.1.0"
-    },
     "npm:get-stream@2.3.1": {
       "buffer": "github:jspm/nodelibs-buffer@0.1.1",
       "object-assign": "npm:object-assign@4.1.1",
       "pinkie-promise": "npm:pinkie-promise@2.0.1",
       "stream": "github:jspm/nodelibs-stream@0.1.0"
-    },
-    "npm:glob@7.1.2": {
-      "assert": "github:jspm/nodelibs-assert@0.1.0",
-      "events": "github:jspm/nodelibs-events@0.1.1",
-      "fs": "github:jspm/nodelibs-fs@0.1.2",
-      "fs.realpath": "npm:fs.realpath@1.0.0",
-      "inflight": "npm:inflight@1.0.6",
-      "inherits": "npm:inherits@2.0.1",
-      "minimatch": "npm:minimatch@3.0.4",
-      "once": "npm:once@1.4.0",
-      "path": "github:jspm/nodelibs-path@0.1.0",
-      "path-is-absolute": "npm:path-is-absolute@1.0.1",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "util": "github:jspm/nodelibs-util@0.1.0"
     },
     "npm:graceful-fs@4.1.11": {
       "assert": "github:jspm/nodelibs-assert@0.1.0",
@@ -369,11 +316,6 @@ System.config({
       "url": "github:jspm/nodelibs-url@0.1.0",
       "util": "github:jspm/nodelibs-util@0.1.0"
     },
-    "npm:inflight@1.0.6": {
-      "once": "npm:once@1.4.0",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "wrappy": "npm:wrappy@1.0.2"
-    },
     "npm:inherits@2.0.1": {
       "util": "github:jspm/nodelibs-util@0.1.0"
     },
@@ -382,9 +324,6 @@ System.config({
     },
     "npm:is-builtin-module@1.0.0": {
       "builtin-modules": "npm:builtin-modules@1.1.1"
-    },
-    "npm:is-fullwidth-code-point@1.0.0": {
-      "number-is-nan": "npm:number-is-nan@1.0.1"
     },
     "npm:isexe@2.0.0": {
       "fs": "github:jspm/nodelibs-fs@0.1.2",
@@ -424,28 +363,17 @@ System.config({
       "util": "github:jspm/nodelibs-util@0.1.0",
       "yallist": "npm:yallist@2.1.2"
     },
-    "npm:materialize-css@0.98.2": {
+    "npm:materialize-css@0.100.1": {
       "hammerjs": "npm:hammerjs@2.0.8",
-      "jquery": "npm:jquery@2.2.4",
-      "node-archiver": "npm:node-archiver@0.3.0"
+      "jquery": "npm:jquery@3.2.1"
     },
     "npm:mem@1.1.0": {
       "mimic-fn": "npm:mimic-fn@1.1.0"
-    },
-    "npm:minimatch@3.0.4": {
-      "brace-expansion": "npm:brace-expansion@1.1.7",
-      "path": "github:jspm/nodelibs-path@0.1.0"
     },
     "npm:mkdirp@0.3.5": {
       "fs": "github:jspm/nodelibs-fs@0.1.2",
       "path": "github:jspm/nodelibs-path@0.1.0",
       "process": "github:jspm/nodelibs-process@0.1.2"
-    },
-    "npm:node-archiver@0.3.0": {
-      "fstream": "npm:fstream@1.0.11",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "tar": "npm:tar@2.2.1",
-      "zlib": "github:jspm/nodelibs-zlib@0.1.0"
     },
     "npm:nopt@2.1.2": {
       "abbrev": "npm:abbrev@1.1.0",
@@ -470,9 +398,6 @@ System.config({
       "path-key": "npm:path-key@2.0.1",
       "process": "github:jspm/nodelibs-process@0.1.2"
     },
-    "npm:once@1.4.0": {
-      "wrappy": "npm:wrappy@1.0.2"
-    },
     "npm:os-locale@2.0.0": {
       "execa": "npm:execa@0.5.1",
       "lcid": "npm:lcid@1.0.0",
@@ -482,10 +407,6 @@ System.config({
     "npm:p-locate@2.0.0": {
       "p-limit": "npm:p-limit@1.1.0"
     },
-    "npm:pako@0.2.9": {
-      "buffer": "github:jspm/nodelibs-buffer@0.1.1",
-      "process": "github:jspm/nodelibs-process@0.1.2"
-    },
     "npm:parse-json@2.2.0": {
       "error-ex": "npm:error-ex@1.3.1"
     },
@@ -494,9 +415,6 @@ System.config({
     },
     "npm:path-exists@3.0.0": {
       "fs": "github:jspm/nodelibs-fs@0.1.2"
-    },
-    "npm:path-is-absolute@1.0.1": {
-      "process": "github:jspm/nodelibs-process@0.1.2"
     },
     "npm:path-key@2.0.1": {
       "process": "github:jspm/nodelibs-process@0.1.2"
@@ -539,16 +457,6 @@ System.config({
       "path": "github:jspm/nodelibs-path@0.1.0",
       "path-type": "npm:path-type@2.0.0"
     },
-    "npm:readable-stream@1.1.14": {
-      "buffer": "github:jspm/nodelibs-buffer@0.1.1",
-      "core-util-is": "npm:core-util-is@1.0.2",
-      "events": "github:jspm/nodelibs-events@0.1.1",
-      "inherits": "npm:inherits@2.0.1",
-      "isarray": "npm:isarray@1.0.0",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "stream-browserify": "npm:stream-browserify@1.0.0",
-      "string_decoder": "npm:string_decoder@1.0.2"
-    },
     "npm:readable-stream@2.2.11": {
       "buffer": "github:jspm/nodelibs-buffer@0.1.1",
       "core-util-is": "npm:core-util-is@1.0.2",
@@ -567,13 +475,6 @@ System.config({
       "path": "github:jspm/nodelibs-path@0.1.0"
     },
     "npm:require-main-filename@1.0.1": {
-      "process": "github:jspm/nodelibs-process@0.1.2"
-    },
-    "npm:rimraf@2.6.1": {
-      "assert": "github:jspm/nodelibs-assert@0.1.0",
-      "fs": "github:jspm/nodelibs-fs@0.1.2",
-      "glob": "npm:glob@7.1.2",
-      "path": "github:jspm/nodelibs-path@0.1.0",
       "process": "github:jspm/nodelibs-process@0.1.2"
     },
     "npm:safe-buffer@5.0.1": {
@@ -615,17 +516,9 @@ System.config({
       "inherits": "npm:inherits@2.0.1",
       "readable-stream": "npm:readable-stream@2.2.11"
     },
-    "npm:string-width@1.0.2": {
-      "code-point-at": "npm:code-point-at@1.1.0",
-      "is-fullwidth-code-point": "npm:is-fullwidth-code-point@2.0.0",
-      "strip-ansi": "npm:strip-ansi@3.0.1"
-    },
     "npm:string-width@2.0.0": {
       "is-fullwidth-code-point": "npm:is-fullwidth-code-point@2.0.0",
       "strip-ansi": "npm:strip-ansi@3.0.1"
-    },
-    "npm:string_decoder@0.10.31": {
-      "buffer": "github:jspm/nodelibs-buffer@0.1.1"
     },
     "npm:string_decoder@1.0.2": {
       "buffer": "github:jspm/nodelibs-buffer@0.1.1",
@@ -633,18 +526,6 @@ System.config({
     },
     "npm:strip-ansi@3.0.1": {
       "ansi-regex": "npm:ansi-regex@2.1.1"
-    },
-    "npm:tar@2.2.1": {
-      "assert": "github:jspm/nodelibs-assert@0.1.0",
-      "block-stream": "npm:block-stream@0.0.9",
-      "buffer": "github:jspm/nodelibs-buffer@0.1.1",
-      "fs": "github:jspm/nodelibs-fs@0.1.2",
-      "fstream": "npm:fstream@1.0.11",
-      "inherits": "npm:inherits@2.0.1",
-      "path": "github:jspm/nodelibs-path@0.1.0",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "stream": "github:jspm/nodelibs-stream@0.1.0",
-      "systemjs-json": "github:systemjs/plugin-json@0.1.2"
     },
     "npm:url@0.10.3": {
       "assert": "github:jspm/nodelibs-assert@0.1.0",