+++ src/collapsible/collapsible.js
@@ -26,6 +26,7 @@ export class MdCollapsible {
   detached() {
     this.attributeManager.removeClasses(['collapsible', 'popout']);
     this.attributeManager.removeAttributes(['data-collapsible']);
+    $(this.element).collapsible('destroy');
   }
 
   refresh() {
@@ -52,5 +53,13 @@ export class MdCollapsible {
 
        handler(targetElement);
      } : null;
+  }
+
+  open(index = 0) {
+    $(this.element).collapsible('open', index);
+  }
+
+  close(index = 0) {
+    $(this.element).collapsible('close', index);
   }
 }