+++ src/select/select.js
@@ -11,6 +11,15 @@ import {DOM} from 'aurelia-pal';
 @customAttribute('md-select')
 export class MdSelect {
   @bindable() disabled = false;
+	@bindable() readonly = false;
+  readonlyChanged() {
+    if (this.readonly) {
+      this.makeReadonly($(this.element).siblings('input')[0]);
+    } else {
+      this.refresh();
+    }
+  }
+
   @bindable() enableOptionObserver = false;
   @bindable() label = '';
   @bindable() showErrortext = true;
@@ -28,9 +37,14 @@ export class MdSelect {
     this.handleBlur = this.handleBlur.bind(this);
     this.log = getLogger('md-select');
     this.bindingEngine = bindingEngine;
+
+    this.handleFocus = this.handleFocus.bind(this);
   }
 
   attached() {
+    if (this.element.classList.contains('browser-default')) {
+      return;
+    }
     let div = $('<div class="input-field"></div>');
     let va = this.element.attributes.getNamedItem('validate');
     if (va) {
@@ -51,6 +65,9 @@ export class MdSelect {
   }
 
   detached() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     let $element = $(this.element);
     $element.off('change', this.handleChangeFromNativeSelect);
     this.observeVisibleDropdownContent(false);
@@ -65,16 +82,25 @@ export class MdSelect {
   }
 
   refresh() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     this.taskQueue.queueTask(() => {
       this.createMaterialSelect(true);
     });
   }
 
   labelChanged(newValue) {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     this.updateLabel();
   }
 
   updateLabel() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     if (this.label) {
       const $label = $(this.element).parent('.select-wrapper').siblings('.md-select-label');
       $label.text(this.label);
@@ -90,6 +116,9 @@ export class MdSelect {
   }
 
   setErrorTextAttribute() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     let input = this.element.parentElement.querySelector('input.select-dropdown');
     if (!input) return;
     this.log.debug('showErrortextChanged: ' + this.showErrortext);
@@ -97,10 +126,16 @@ export class MdSelect {
   }
 
   notifyBindingEngine() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     this.log.debug('selectedOptions changed', arguments);
   }
 
   handleChangeFromNativeSelect() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     if (!this._suspendUpdate) {
       this.log.debug('handleChangeFromNativeSelect', this.element.value, $(this.element).val());
       this._suspendUpdate = true;
@@ -110,6 +145,9 @@ export class MdSelect {
   }
 
   handleChangeFromViewModel(newValue) {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     this.log.debug('handleChangeFromViewModel', newValue, $(this.element).val());
     if (!this._suspendUpdate) {
       this.createMaterialSelect(false);
@@ -117,6 +155,9 @@ export class MdSelect {
   }
 
   toggleControl(disable) {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     let $wrapper = $(this.element).parent('.select-wrapper');
     if ($wrapper.length > 0) {
       if (disable) {
@@ -127,25 +168,53 @@ export class MdSelect {
         $('.caret', $wrapper).removeClass('disabled');
         $('input.select-dropdown', $wrapper).attr('disabled', null);
         $wrapper.attr('disabled', null);
-        $('.select-dropdown', $wrapper).dropdown({'hover': false, 'closeOnClick': false});
       }
     }
   }
 
   createMaterialSelect(destroy) {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     this.observeVisibleDropdownContent(false);
     this.observeOptions(false);
+    let input = $(this.element).siblings("input");
+    let isValid = input.hasClass("valid");
+    let isInvalid = input.hasClass("invalid");    
     if (destroy) {
       $(this.element).material_select('destroy');
     }
     $(this.element).material_select();
+    input = $(this.element).siblings("input");
+    if (isValid) {
+      input.addClass("valid");
+    }
+    if (isInvalid) {
+      input.addClass("invalid");
+    }    
     this.toggleControl(this.disabled);
     this.observeVisibleDropdownContent(true);
     this.observeOptions(true);
     this.setErrorTextAttribute();
+    if (this.readonly) {
+      this.makeReadonly(input[0]);
+    }
+  }
+
+  makeReadonly(input) {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
+    $(input).off('click');
+    $(input).off('focus');
+    $(input).off('keydown');
+    $(input).off('open');
   }
 
   observeVisibleDropdownContent(attach) {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     if (attach) {
       if (!this.dropdownMutationObserver) {
         this.dropdownMutationObserver = DOM.createMutationObserver(mutations => {
@@ -158,6 +227,8 @@ export class MdSelect {
           if (isHidden) {
             this.dropdownMutationObserver.takeRecords();
             this.handleBlur();
+          } else {
+            this.handleFocus();
           }
         });
       }
@@ -174,6 +245,9 @@ export class MdSelect {
   }
 
   observeOptions(attach) {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     if (getBooleanFromAttributeValue(this.enableOptionObserver)) {
       if (attach) {
         if (!this.optionsMutationObserver) {
@@ -197,6 +271,9 @@ export class MdSelect {
   }
 
   open() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     $(this.element).siblings('input.select-dropdown').trigger('focus');
   }
 
@@ -212,12 +289,30 @@ export class MdSelect {
   _taskqueueRunning = false;
 
   handleBlur() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
     if (this._taskqueueRunning) return;
     this._taskqueueRunning = true;
     this.taskQueue.queueTask(() => {
       this.log.debug('fire blur event');
       fireEvent(this.element, 'blur');
       this._taskqueueRunning = false;
+
+      if (this.label) {
+        const $label = $(this.element).parent('.select-wrapper').siblings('.md-select-label');
+        $label.removeClass('md-focused');
+      }
     });
+  }
+
+  handleFocus() {
+    if ((this.element).classList.contains('browser-default')) {
+      return;
+    }
+    if (this.label) {
+      const $label = $(this.element).parent('.select-wrapper').siblings('.md-select-label');
+      $label.addClass('md-focused');
+    }
   }
 }