+++ src/sidenav/sidenav-collapse.js
@@ -1,41 +1,38 @@
 import { bindable, customAttribute } from 'aurelia-templating';
-import { ObserverLocator } from 'aurelia-binding';
+import { getLogger } from 'aurelia-logging';
 import { inject } from 'aurelia-dependency-injection';
 import { getBooleanFromAttributeValue } from '../common/attributes';
-import { getLogger } from 'aurelia-logging';
+import { fireMaterializeEvent } from '../common/events';
 
 @customAttribute('md-sidenav-collapse')
-@inject(Element, ObserverLocator)
+@inject(Element)
 export class MdSidenavCollapse {
   @bindable() ref;
-  constructor(element, observerLocator) {
+  constructor(element) {
     this.element = element;
-    this.observerLocator = observerLocator;
     this.log = getLogger('md-sidenav-collapse');
   }
 
   attached() {
     this.ref.whenAttached.then(() => {
-      // this.widthSubscription = this.observerLocator.getObserver(this.ref, 'mdWidth').subscribe(this.widthChanged.bind(this));
-      // this.fixedSubscription = this.observerLocator.getObserver(this.ref, 'fixed').subscribe(this.fixedChanged.bind(this));
-
       const closeOnClick = this.ref.mdFixed && window.innerWidth > 992 ? false : getBooleanFromAttributeValue(this.ref.mdCloseOnClick);
+
+      this.onHide = this.onHide.bind(this);
+      this.onShow = this.onShow.bind(this);
 
       this.element.setAttribute('data-activates', this.ref.controlId);
       let sideNavConfig = {
         edge: this.ref.mdEdge || 'left',
-        // closeOnClick: (this.ref.mdFixed ? false : getBooleanFromAttributeValue(this.ref.mdCloseOnClick)),
         closeOnClick,
-        menuWidth: parseInt(this.ref.mdWidth, 10)
+        menuWidth: parseInt(this.ref.mdWidth, 10),
+        onClose: this.onHide,
+        onOpen: this.onShow
       };
-      // this.log.debug('sideNavConfig:', sideNavConfig);
       $(this.element).sideNav(sideNavConfig);
     });
   }
 
-  detached() {
-    // this.widthSubscription.unsubscribe();
-  }
+  detached() { }
 
   show() {
     $(this.element).sideNav('show');
@@ -45,21 +42,11 @@ export class MdSidenavCollapse {
     $(this.element).sideNav('hide');
   }
 
-  // fixedChanged() {
-  //   this.log.debug('fixedChanged');
-  //   $(this.element).sideNav({
-  //     edge: this.ref.edge || 'left',
-  //     closeOnClick: this.ref.closeOnClick,
-  //     menuWidth: parseInt(this.ref.mdWidth, 10)
-  //   });
-  // }
-  //
-  // widthChanged() {
-  //   this.log.debug('widthChanged');
-  //   $(this.element).sideNav({
-  //     edge: this.ref.edge || 'left',
-  //     closeOnClick: this.ref.closeOnClick,
-  //     menuWidth: parseInt(this.ref.mdWidth, 10)
-  //   });
-  // }
+  onShow(el) {
+    fireMaterializeEvent(this.ref.element, 'sidenav-show');
+  }
+
+  onHide(el) {
+    fireMaterializeEvent(this.ref.element, 'sidenav-hide');
+  }
 }