+++ src/toast/toastService.js
@@ -1,8 +1,12 @@
 export class MdToastService {
+  removeAll() {
+    Materialize.Toast.removeAll();
+  }
+
   show(message, displayLength, className?) {
     return new Promise((resolve, reject) => {
-      Materialize.toast(message, displayLength, className, () => {
-        resolve();
+      const toastInstance = Materialize.toast(message, displayLength, className, () => {
+        resolve(toastInstance);
       });
     });
   }